# docs/source/reference/api-restapi/openapi.yml
openapi: 3.0.1
info:
  title: Radar REST API
  version: 0.0.0
servers:
- url: /
...
paths:
...
  /api/v1/groups/:
    get:
      tags:
      - Groups
      summary: Gets a list of all Group resources
      operationId: get_group_endpoint
      parameters:
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPage'
  /api/v1/groups/{id}:
    get:
      tags:
      - Groups
      summary: Gets a Group resource
      operationId: get_group_id_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the Group resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group-dump'
...
components:
  schemas:
...
    GroupRef:
      type: object
      properties:
        id:
          type: integer
          description: ID for the Group resource.
        name:
          type: string
          description: Name of the Group resource.
        url:
          type: string
          description: URL for accessing the full Group resource.
    UserRef-dump:
      type: object
      properties:
        id:
          type: integer
          description: ID for the User resource.
        username:
          type: string
          description: Username of the User resource.
        url:
          type: string
          description: URL for accessing the full User resource.
...
    GroupPage:
      type: object
      properties:
        index:
          type: integer
          description: Index of the current page.
        isComplete:
          type: boolean
          description: Boolean indicating if more data is available.
        totalNumResults:
          type: integer
          description: Total number of results.
        first:
          type: string
          description: URL to first page in results set.
        next:
          type: string
          description: URL to next page in results set.
        prev:
          type: string
          description: URL to last page in results set.
        data:
          type: array
          items:
            type: object
            description: List of Group resources in the current page.
            allOf:
            - $ref: '#/components/schemas/Group-dump'
    Group-dump:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          description: Name of the Group.
        id:
          type: integer
          description: ID for the Group resource.
        user:
          type: object
          description: User that created the Group resource.
          allOf:
          - $ref: '#/components/schemas/UserRef-dump'
        members:
          type: array
          items:
            type: object
            description: A list of GroupMembers in a Group.
            allOf:
            - $ref: '#/components/schemas/GroupMember-dump'
        createdOn:
          type: string
          description: Timestamp when the Group resource was created.
          format: date-time
        lastModifiedOn:
          type: string
          description: Timestamp when the Group resource was last modified.
          format: date-time
    GroupMember-dump:
      type: object
      properties:
        user:
          type: object
          description: User that is a member of the Group.
          allOf:
          - $ref: '#/components/schemas/UserRef-dump'
        group:
          type: object
          description: The Group of which the User is a member.
          allOf:
          - $ref: '#/components/schemas/GroupRef'
        permissions:
          type: object
          description: The Group Permissions for this User.
          allOf:
          - $ref: '#/components/schemas/Generated-dump'
    Generated-dump:
      type: object
      properties:
        read:
          type: boolean
          description: Permission for member to read Group.
        write:
          type: boolean
          description: Permission for member to modify Group.
        shareRead:
          type: boolean
          description: Permission for member to share read-only Group.
        shareWrite:
          type: boolean
          description: Permission for member to share read+write Group.
        owner:
          type: boolean
          description: Flag for if the member is a Group owner.
        admin:
          type: boolean
          description: Flag for if the member is a Group admin.
...